Add support for TOAST chunk_id type in binary upgrades
This commit adds a new function, which would set the type of a chunk_id
attribute for a TOAST table across upgrades.  This piece currently works
only with chunk_id = OIDOID, but it is required in a follow-up patch
where support for chunk_id = OID8OID is supported on top of the existing
one.