Add relation option toast_value_type
This relation option gives the possibility to define the attribute type
that can be used for chunk_id in a TOAST table when it is created
initially.  This parameter has no effect if a TOAST table exists, even
after it is modified later on, even on rewrites.

This can be set only to "oid" currently, and will be expanded with a
second mode later.

Note: perhaps it would make sense to introduce that only when support
for 8-byte OID values are added to TOAST, the split is here to ease
review.