Preserve index stats during ALTER TABLE ... TYPE ...
During ALTER TABLE ... TYPE ... on an indexed column, a new index is created and
the old one is dropped. Currently this causes the statistics such as idx_scan,
last_idx_scan, and related counters to be lost.

We can not use pgstat_copy_relation_stats() because the old index is dropped
before the new one is created, so this commit saves the old PgStat_StatTabEntry
and restores it in the new index. This is done by adding a pointer to a hash table
in AlteredTableInfo where the hash key is the partition table Oid. This hash table
stores a list of all the information needed to restore the stats in the right
index.

The stats are saved in ATPostAlterTypeParse() (before the old index is dropped)
and restored in ATExecAddIndex() once the new index is created.

Note that pending statistics (if any) are not preserved, only the
accumulated stats from previous transactions. This is acceptable since the
accumulated stats represent the historical usage patterns we want to maintain.