CREATE VARIABLE, DROP VARIABLE
Introduce simple (non transactional) implementation of basic DDL statements.
Because only temporary variables are allowed, we don't need a dedicated catalog.

This is the most reduced version of declared session variables implementation.
Main goal is introduction of basic functionality and reducing the size of the patch
to maximum. This patch is designed to be possible to move to catalog based
implementation without compatibility break.

Some details about different implementations of session variables

https://wiki.postgresql.org/wiki/Implementation_of_declarative_catalog_session_variables